Fix a bug for Vala when two tasks share build files.
If two tasks, let's say a shared and a static library, share the same
sources *.vala, they will generate the same *.c files, but since it's
part of the valac task to move these files, they won't be available for
the second task.

This is a naive fix that simply replaces shutil.move for shutil.copy.
A good fix would probably generate the files at the right locations and
simply avoiding the move call.